+++ packages/app/src/index.ts
@@ -1,3 +1,11 @@
 export { initializeAppModules, appModules } from './modules';
-export type { AppManifest, AppModules, AppModulesInstance, AppConfigurator } from './modules';
+
+export type {
+    AppManifest,
+    AppModules,
+    AppModulesInstance,
+    AppConfigurator,
+    AppConfigCallback,
+} from './modules';
+
 export { default } from './modules';