+++ packages/framework-react/src/index.tsx
@@ -4,6 +4,7 @@
  */
 
 export type { Fusion, FusionConfigurator } from '@equinor/fusion-framework';
+export { initFusion } from '@equinor/fusion-framework';
 
 export { createFrameworkProvider } from './create-framework-provider';
 export { FrameworkProvider as default, FrameworkProvider } from './context';