+++ packages/react-app/src/http/index.ts
@@ -1,7 +1,1 @@
-export {
-    IHttpClient,
-    HttpClientMsal,
-    FetchRequestInit,
-} from '@equinor/fusion-framework-module-http';
-
-export { useHttpClient } from './useHttpClient';
+export * from '@equinor/fusion-framework-react/http';