+++ packages/app/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.2.3 (2022-09-05)
+
+**Note:** Version bump only for package @equinor/fusion-framework-app
+
+
+
+
+
+## 0.2.2 (2022-09-05)
+
+**Note:** Version bump only for package @equinor/fusion-framework-app
+
+
+
+
+
+## 0.2.1 (2022-09-01)
+
+**Note:** Version bump only for package @equinor/fusion-framework-app
+
+
+
+
+
 ## [0.2.0](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-app@0.1.2...@equinor/fusion-framework-app@0.2.0) (2022-08-29)