+++ packages/module-msal/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.4.2 (2022-09-05)
+
+**Note:** Version bump only for package @equinor/fusion-framework-module-msal
+
+
+
+
+
+## 0.4.1 (2022-09-05)
+
+**Note:** Version bump only for package @equinor/fusion-framework-module-msal
+
+
+
+
+
 ## 0.4.0 (2022-08-29)