+++ packages/react-app/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.3](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-react-app@0.7.2...@equinor/fusion-framework-react-app@0.7.3) (2022-09-05)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-app
+
+
+
+
+
+## [0.7.2](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-react-app@0.7.1...@equinor/fusion-framework-react-app@0.7.2) (2022-09-05)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-app
+
+
+
+
+
+## 0.7.1 (2022-09-05)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-app
+
+
+
+
+
+## [0.7.0](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-react-app@0.6.1...@equinor/fusion-framework-react-app@0.7.0) (2022-09-05)
+
+
+### Features
+
+* **react-app:** expose http client ([91b9930](https://github.com/equinor/fusion-framework/commit/91b9930f404772bd58ce043b6987aaffc8324654))
+
+
+
+## 0.6.1 (2022-09-01)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-app
+
+
+
+
+
 ## [0.6.0](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-react-app@0.5.1...@equinor/fusion-framework-react-app@0.6.0) (2022-08-29)